requirements.txt
Django>=1.8, <1.9
amqp==1.4.5
anyjson==0.3.3
billiard==3.3.0.18
boto==2.31.1
celery==3.1.13
certifi==14.05.14
coverage==3.7.1
dj-database-url==0.3.0
dj-static==0.0.6
django-celery==3.1.16
django-filter==0.10.0
django-nose==1.4
djangorestframework==3.1.2
drf-any-permissions==0.0.1
flake8==2.2.5
gunicorn==19.0.0
kombu==3.0.21
mccabe==0.2.1
mock==1.0.1
nose==1.3.4
oauthlib==0.6.3
pep8==1.5.7
psycopg2==2.5.3
pycleaner==1.1.1
pyflakes==0.8.1
python-json-logger==0.1.2
python-openid==2.2.5
python-social-auth==0.1.26
pytz==2014.4
requests==2.6.0
requests-oauthlib==0.4.1
six==1.7.3
splunk-handler==1.1.2
static3==0.5.1
stripe==1.19.0
wsgiref==0.1.2
